fix build failure on 32-bit machine

+++ source/s3v2/s3.go
@@ -46,7 +46,7 @@ type S3File struct {
 	readOpened     bool
 	fileSize       int64
 	socket         io.ReadCloser
-	minRequestSize int
+	minRequestSize int64
 
 	lock       sync.RWMutex
 	err        error
@@ -189,7 +189,7 @@ func (s *S3File) Read(p []byte) (n int, err error) {
 	}()
 
 	if s.socket == nil {
-		err = s.openSocket(len(p))
+		err = s.openSocket(int64(len(p)))
 		if err != nil {
 			return 0, err
 		}
@@ -210,7 +210,7 @@ func (s *S3File) Read(p []byte) (n int, err error) {
 
 // openSocket issues a new GetObject request to retrieve the next chunk of data from the
 // object.
-func (s *S3File) openSocket(numBytes int) error {
+func (s *S3File) openSocket(numBytes int64) error {
 	if numBytes < s.minRequestSize {
 		numBytes = s.minRequestSize
 	}
@@ -402,7 +402,7 @@ func (s *S3File) openRead() error {
 }
 
 // getBytesRange returns the range request header string
-func (s *S3File) getBytesRange(numBytes int) string {
+func (s *S3File) getBytesRange(numBytes int64) string {
 	var (
 		byteRange string
 		begin     int64
@@ -474,7 +474,7 @@ func NewS3FileReaderWithParams(ctx context.Context, params S3FileReaderParams) (
 		s3Client = s3.NewFromConfig(getConfig())
 	}
 
-	minRequestSize := params.MinRequestSize
+	minRequestSize := int64(params.MinRequestSize)
 	if minRequestSize == 0 {
 		minRequestSize = defaultMinRequestSize
 	}
